generate encrypted disk

+++ misc/create-disk-encrypted.sh
@@ -8,14 +8,16 @@ PASS=$3
 MNTDIR=$(mktemp -d)
 SIZE=100M
 
-echo "input dir=$INPUTDIR"
-echo "disk=$DISK"
-echo "mount dir=$MNTDIR"
+echo "INPUTDIR=$INPUTDIR"
+echo "DISK=$DISK"
+echo "MNTDIR=$MNTDIR"
+DISKFILE=$(basename $DISK)
+echo "DISKFILE=$DISKFILE"
 
-rm -i $DISK
+[ -f $DISK ] && rm -i $DISK
 
 [ ! -d $INPUTDIR ] && echo "Error: input directory not found!" && exit 0
-[ -f $DISK ] && echo "Error: disk file already exists!" && exit 0
+# [ -f $DISK ] && echo "Error: disk file already exists!" && exit 0
 
 
 dd if=/dev/zero of=$DISK bs=$SIZE count=1   # empty floppy disk image
@@ -29,8 +31,8 @@ sudo mount -o loop,uid=$UID,gid=$UID $DISK $MNTDIR
 ls -ld $MNTDIR
 
 export PASS
-( cd $INPUTDIR ; tar cvf $MNTDIR/$DISK.tar * )
-openssl enc -in $MNTDIR/$DISK.tar -out $MNTDIR/$DISK.tar.enc -pass env:PASS
+( cd $INPUTDIR ; tar cvf $MNTDIR/$DISKFILE.tar * )
+openssl enc -in $MNTDIR/$DISKFILE.tar -out $MNTDIR/$DISKFILE.tar.enc -pass env:PASS
 # rm -f $MNTDIR/$DISK.tar
 sudo umount $MNTDIR
